--- src/main/java/com/ohgiraffers/adregamdi/category/command/domain/aggregate/entity/test.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.category.command.domain.aggregate.entity;
-
-public class test {
-}
