+++ src/main/java/com/ohgiraffers/adregamdi/place/command/domain/repository/DongRepository.java
@@ -1,6 +1,6 @@
 package com.ohgiraffers.adregamdi.place.command.domain.repository;
 
-import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.Dong;
+import com.ohgiraffers.adregamdi.cityanddong.command.domain.aggregate.entity.Dong;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DongRepository extends JpaRepository<Dong, Long> {
