+++ src/main/java/com/ohgiraffers/adregamdi/place/command/domain/service/PlaceAPIService.java
@@ -1,6 +1,8 @@
 package com.ohgiraffers.adregamdi.place.command.domain.service;
 
 import com.ohgiraffers.adregamdi.category.command.domain.aggregate.entity.Category;
+import com.ohgiraffers.adregamdi.cityanddong.command.domain.aggregate.entity.City;
+import com.ohgiraffers.adregamdi.cityanddong.command.domain.aggregate.entity.Dong;
 import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.*;
 import com.ohgiraffers.adregamdi.tag.command.domain.aggregate.entity.Tag;
