+++ src/main/java/com/ohgiraffers/adregamdi/place/query/infra/repository/CityQueryRepository.java
@@ -1,6 +1,6 @@
 package com.ohgiraffers.adregamdi.place.query.infra.repository;
 
-import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.City;
+import com.ohgiraffers.adregamdi.cityanddong.command.domain.aggregate.entity.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CityQueryRepository extends JpaRepository<City, Integer> {
