--- src/main/java/com/ohgiraffers/adregamdi/tag/command/domain/aggregate/entity/test.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.tag.command.domain.aggregate.entity;
-
-public class test {
-}
