--- src/main/java/com/ohgiraffers/adregamdi/tagplace/command/application/controller/test.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.tagplace.command.application.controller;
-
-public class test {
-}
