--- src/main/java/com/ohgiraffers/adregamdi/tagplace/command/application/dto/test.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.tagplace.command.application.dto;
-
-public class test {
-}
