--- src/main/java/com/ohgiraffers/adregamdi/tagplace/command/application/service/test.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.tagplace.command.application.service;
-
-public class test {
-}
