--- src/main/java/com/ohgiraffers/adregamdi/tagplace/command/domain/aggregate/entity/PlaceTags.java
@@ -1,31 +1,0 @@
-package com.ohgiraffers.adregamdi.tagplace.command.domain.aggregate.entity;
-
-import com.ohgiraffers.adregamdi.tagplace.command.domain.aggregate.vo.PlaceVO;
-import com.ohgiraffers.adregamdi.tagplace.command.domain.aggregate.vo.TagVO;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Getter
-@NoArgsConstructor
-@Entity
-@Table(name = "TBL_PLACETAGS")
-public class PlaceTags {
-
-    @Id
-    @Column(name = "place_tags_no")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long placeTagsNo;
-
-    @Embedded
-    private PlaceVO placeVO;
-
-    @Embedded
-    private TagVO tagVO;
-
-    public PlaceTags(PlaceVO placeVO, TagVO tagVO) {
-        this.placeVO = placeVO;
-        this.tagVO = tagVO;
-    }
-}
