--- src/main/java/com/ohgiraffers/adregamdi/tagplace/command/domain/aggregate/vo/test.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.tagplace.command.domain.aggregate.vo;
-
-public class test {
-}
