--- src/main/java/com/ohgiraffers/adregamdi/tagplace/command/domain/exception/test.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.tagplace.command.domain.exception;
-
-public class test {
-}
