--- src/main/java/com/ohgiraffers/adregamdi/tagplace/command/domain/repository/test.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.tagplace.command.domain.repository;
-
-public class test {
-}
