--- src/main/java/com/ohgiraffers/adregamdi/tagplace/command/infra/repository/test.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.tagplace.command.infra.repository;
-
-public class test {
-}
