--- src/main/java/com/ohgiraffers/adregamdi/tagplace/query/application/controller/test.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.tagplace.query.application.controller;
-
-public class test {
-}
