--- src/main/java/com/ohgiraffers/adregamdi/tagplace/query/application/dto/test.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.tagplace.query.application.dto;
-
-public class test {
-}
