--- src/main/java/com/ohgiraffers/adregamdi/tagplace/query/application/service/test.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.tagplace.query.application.service;
-
-public class test {
-}
