--- src/main/java/com/ohgiraffers/adregamdi/tagplace/query/infra/repository/test.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.tagplace.query.infra.repository;
-
-public class test {
-}
