--- src/main/java/com/ohgiraffers/adregamdi/tagplace/query/infra/service/test.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.tagplace.query.infra.service;
-
-public class test {
-}
