+++ src/test/java/com/ohgiraffers/adregamdi/place/command/domain/service/DataDomainServiceTest.java
@@ -1,18 +1,14 @@
 package com.ohgiraffers.adregamdi.place.command.domain.service;
 
-import org.json.JSONException;
+import com.ohgiraffers.adregamdi.data.command.domain.service.DataDomainService;
 import org.json.simple.JSONObject;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 
 import java.util.Arrays;
 import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 
 class DataDomainServiceTest {
