+++ src/test/java/com/ohgiraffers/adregamdi/place/query/application/service/DataQueryServiceTest.java
@@ -1,9 +1,9 @@
 package com.ohgiraffers.adregamdi.place.query.application.service;
 
-import com.ohgiraffers.adregamdi.place.command.application.service.DataService;
+import com.ohgiraffers.adregamdi.data.command.application.service.DataService;
 import com.ohgiraffers.adregamdi.category.command.domain.aggregate.entity.Category;
-import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.City;
-import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.Dong;
+import com.ohgiraffers.adregamdi.cityanddong.command.domain.aggregate.entity.City;
+import com.ohgiraffers.adregamdi.cityanddong.command.domain.aggregate.entity.Dong;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
